changed id data type from int to long

+++ src/main/java/com/labC/libraryCheckoutSystem/model/User.java
@@ -10,11 +10,11 @@ public class User {
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private int id;
+    private long id;
     private String name;
     private String email;
 
-    public User(int id, String name, String email) {
+    public User(long id, String name, String email) {
         this.id = id;
         this.name = name;
         this.email = email;
@@ -22,7 +22,7 @@ public class User {
 
     public User(){}
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
